pyproject.toml
[project]
name = "torretray-bot"
version = "0.1.0"
description = "Telegram bot for TorreTray"
requires-python = ">=3.11,<3.15"
dependencies = [
  "python-telegram-bot>=21.11,<22.0",
  "httpx>=0.27.0,<1.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
]

[project.scripts]
torretray-bot = "torretray_bot.app:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["torretray_bot*"]
